+++ tests/run.clj
@@ -23,6 +23,7 @@
 (require "tests/lazy_seq_ns_scope_test")
 (require "tests/jvm_statics_test")
 (require "tests/clojure_version_test")
+(require "tests/inst_test")
 (require "tests/macro_test")
 (require "tests/error_test")
 (require "tests/atom_test")